requirements.txt
streamlit==1.45.1
openai==1.30.0
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
faiss-cpu==1.9.0.post1
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
scikit-learn==1.5.0
python-dotenv==1.0.1
tiktoken==0.7.0
pydantic==2.7.1
httpx==0.27.0
sentence-transformers==3.0.1
torch==2.3.0